+++ CHANGELOG.md
@@ -7,12 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- feat: add the possibility for the user to pass additional metadata to the compute plan metadata (#161)
+
 ### Fixed
 
 - Force the reinstallation of connect-tools in the Docker image, necessary for the editable mode (#156)
 
 ### Changed
-
 
 - BREAKING CHANGE: the default value of ``drop_last`` in the ``NpIndexGenerator`` is now False (#142)
 - BREAKING CHANGE: the index generator is now required when implementing a strategy (#142)
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         )
     # ...
   ```
+
 - The user can now initialize his `TorchAlgo` function with custom parameters (only primitive types are supported) (#130):
 
     ```python